feat(cache): add redis caching for like, comment, and post

+++ src/main/java/SocialNetwork/SocialNetwork/repositories/CommentRepository.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CommentRepository extends JpaRepository<Comment,Integer> {
+public interface CommentRepository extends JpaRepository<Comment,Long> {
     List<Comment> findAllByPost(Post post);
     Comment findByUserAndPostAndId(User user, Post post, Long commentId);
     @Query("SELECT COUNT(c) FROM Comment c WHERE c.post = :post")
